Create lamp.sh
LAMP was more difficult it would also add the linux == "L" in the word LAMP. Linux partitioning automation from source would be a bit difficult, also I will need to add the unix version in the next push, the difference between linux and unix is small so it won't take long until it would be finished. Documentation README.md will be added after the unix version.